putting theme switcher in footer
