+++ components/custom-layout-elements/footer.tsx
@@ -1,6 +1,49 @@
 import Link from 'next/link';
+import { Switch } from '@mui/material';
+import { Moon, Sun } from 'lucide-react';
+import { Label } from '@/components/ui/label';
+import { useTheme } from 'next-themes';
+import { createTheme } from '@mui/material/styles';
+import { useTheme as useMuiTheme } from '@mui/material/styles';
+import { useState, useEffect } from 'react';
+
+declare module '@mui/material/styles' {
+    interface Palette {
+        custom: Palette['primary'];
+    }
+
+    interface PaletteOptions {
+        custom?: PaletteOptions['primary'];
+    }
+}
+
+declare module '@mui/material/Switch' {
+    interface SwitchPropsColorOverrides {
+        custom: true;
+    }
+}
 
 export function Footer() {
+    const [mounted, setMounted] = useState(false);
+    const { theme, setTheme } = useTheme();
+
+    useEffect(() => {
+        setMounted(true);
+    }, []);
+
+    const muiTheme = createTheme({
+        palette: {
+            primary: {
+                main: '#3C2D63',
+                light: '#D0BCFF',
+                dark: '#251a3d',
+                contrastText: ' #EADDFF',
+            },
+        },
+    });
+
+    const usePurpleTheme = useMuiTheme();
+
     const year = new Date().getFullYear();
     return (
         <footer className="items-center h-30 w-full bg-midnight-purple dark:bg-ebena-lavender text-barely-lilac dark:text-pompaca-purple px-4 py-4 mt-auto inset-shadow-sm inset-shadow-gray-500">
@@ -27,42 +70,74 @@ export function Footer() {
                         </a>
                     </span>
                 </div>
-                <div className="flex items-center gap-4">
-                    <Link href="/terms" className="hover:underline">
-                        Terms of Service
-                    </Link>
-                    <span>|</span>
-                    <Link href="/privacy" className="hover:underline">
-                        Privacy Policy
-                    </Link>
-                    <span>|</span>
-                    <a
-                        href="https://github.com/rio-codes/tfo-creaturetracker"
-                        className="hover:underline"
-                    >
-                        Github
-                    </a>
-                    <span>|</span>
-                    <a
-                        href="https://www.patreon.com/cw/tfoct"
-                        className="hover:underline"
-                    >
-                        Patreon
-                    </a>
-                    <span>|</span>
-                    <a
-                        href="https://discord.gg/PMtE3jrXYR"
-                        className="hover:underline"
-                    >
-                        Discord
-                    </a>
-                    <span>|</span>
-                    <a
-                        href="https://finaloutpost.net/lab/lyricism"
-                        className="hover:underline"
-                    >
-                        My Lab
-                    </a>
+                <div className="flex-col justify-items-end">
+                    <div className="flex items-center gap-4">
+                        <Link href="/terms" className="hover:underline">
+                            Terms of Service
+                        </Link>
+                        <span>|</span>
+                        <Link href="/privacy" className="hover:underline">
+                            Privacy Policy
+                        </Link>
+                        <span>|</span>
+                        <a
+                            href="https://github.com/rio-codes/tfo-creaturetracker"
+                            className="hover:underline"
+                        >
+                            Github
+                        </a>
+                        <span>|</span>
+                        <a
+                            href="https://www.patreon.com/cw/tfoct"
+                            className="hover:underline"
+                        >
+                            Patreon
+                        </a>
+                        <span>|</span>
+                        <a
+                            href="https://discord.gg/PMtE3jrXYR"
+                            className="hover:underline"
+                        >
+                            Discord
+                        </a>
+                        <span>|</span>
+                        <a
+                            href="https://finaloutpost.net/lab/lyricism"
+                            className="hover:underline"
+                        >
+                            My Lab
+                        </a>
+                    </div>
+                    {mounted && (
+                        <div className="flex w-25 items-center justify-end py-2">
+                            <Label
+                                htmlFor="theme-switch"
+                                className="dark:text-pompaca-purple text-barely-lilac"
+                            >
+                                <Sun className="h-4 w-4" />
+                            </Label>
+                            <Switch
+                                id="theme-switch"
+                                defaultValue={
+                                    theme === 'system' ? undefined : theme
+                                }
+                                checked={theme === 'dark'}
+                                onChange={() =>
+                                    setTheme(
+                                        theme === 'dark' ? 'light' : 'dark'
+                                    )
+                                }
+                                color="custom"
+                                size="medium"
+                            />
+                            <Label
+                                htmlFor="theme-switch"
+                                className="dark:text-pompaca-purple  text-barely-lilac"
+                            >
+                                <Moon className="h-4 w-4" />
+                            </Label>
+                        </div>
+                    )}
                 </div>
             </div>
         </footer>
